scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pymongo==4.6.2
scikit-learn==1.4.1.post1
pandas==2.2.1
python-dotenv==1.0.1
